fix: kineokit tests, model exporting in main package
